tests scripts fixes for laser

+++ tests/make_tests.sh
@@ -1135,14 +1135,16 @@ do_hash_test ()
 
  # for text files, we remove potential CR chars
  # to prevent having different hashes on different platforms
- if [ -n "$(file -b $1 | grep text)" ] ||  [ ${1: -4} == ".lsr" ] ; then
+ if [ -n "$(file -b $1 | grep text)" ] ||  [ ${1: -4} == ".lsr" ] ||  [ ${1: -4} == ".svg" ] ; then
   file_to_hash="to_hash_$(basename $1)"
-  tr -d '\r' <  "$1" > "$file_to_hash"
+  if [ -f $1 ]; then
+    tr -d '\r' <  "$1" > "$file_to_hash"
+  fi
  fi
 
  $MP4BOX -hash -std $file_to_hash > $test_hash 2>> $log_subtest
 
- if [ "$file_to_hash" != "$1" ]; then
+ if [ "$file_to_hash" != "$1" ] && [ -f "$file_to_hash" ]; then
   rm "$file_to_hash"
  fi
